Conflicts:
	main/src/main/java/com/bloatit/model/Comment.java
	main/src/main/java/com/bloatit/model/Feature.java
	main/src/main/java/com/bloatit/model/Release.java

+++ main/src/main/java/com/bloatit/model/FeatureImplementation.java
@@ -23,9 +23,11 @@ import java.util.Locale;
 import com.bloatit.common.Log;
 import com.bloatit.data.DaoComment;
 import com.bloatit.data.DaoContribution;
+import com.bloatit.data.DaoContribution.State;
 import com.bloatit.data.DaoDescription;
 import com.bloatit.data.DaoFeature;
 import com.bloatit.data.DaoFeature.FeatureState;
+import com.bloatit.data.DaoMember.Role;
 import com.bloatit.data.DaoOffer;
 import com.bloatit.data.DaoTeamRight.UserTeamRight;
 import com.bloatit.data.exceptions.NotEnoughMoneyException;
@@ -136,6 +138,51 @@ public final class FeatureImplementation extends Kudosable<DaoFeature> implement
     @Override
     public boolean canAccessOffer(final Action action) {
         return canAccess(new RgtFeature.Offer(), action);
+    }
+
+    /**
+     * Indicates whether the feature can be modified or not
+     * <p>
+     * A feature can be modified when :
+     * <li>The connected user is admin</li>
+     * <li>OR the connected user is the author of the feature</li>
+     * <li>AND the feature has no contributions (except from the author of the
+     * feature)</li>
+     * <li>AND the feature has no offer (except from the authoer of the feature)
+     * </li>
+     * </p>
+     */
+    @Override
+    public boolean canModify() {
+        if (!AuthToken.isAuthenticated()) {
+            return false;
+        }
+
+        if (AuthToken.getMember().getRole() == Role.ADMIN) {
+            return true;
+        }
+
+        if (!AuthToken.getMember().equals(getMember())) {
+            return false;
+        }
+
+        if (!(getFeatureState() == FeatureState.PENDING || getFeatureState() == FeatureState.PREPARING)) {
+            return false;
+        }
+
+        for (Contribution c : getContributions(false)) {
+            if (!c.getMember().equals(getMember())) {
+                return false;
+            }
+        }
+
+        for (Offer o : getOffers()) {
+            if (!o.getMember().equals(getMember())) {
+                return false;
+            }
+        }
+
+        return true;
     }
 
     // /////////////////////////////////////////////////////////////////////////////////////////
@@ -296,16 +343,10 @@ public final class FeatureImplementation extends Kudosable<DaoFeature> implement
     }
 
     @Override
-    public void delete() throws UnauthorizedOperationException {
-        if (isDeleted()) {
-            return;
+    protected void delete(boolean delOrder) throws UnauthorizedOperationException {
+        if (delOrder) {
+            this.setFeatureState(FeatureState.DISCARDED);
         }
-
-        if (!getRights().hasAdminUserPrivilege()) {
-            throw new UnauthorizedOperationException(SpecialCode.ADMIN_ONLY);
-        }
-
-        this.setFeatureState(FeatureState.DISCARDED);
 
         // We delete every components of the feature (comments, offers,
         // translations ...)
@@ -315,23 +356,22 @@ public final class FeatureImplementation extends Kudosable<DaoFeature> implement
         // on the web site (for example on the account page) which is the
         // behavior we want.
 
-        super.delete();
+        super.delete(delOrder);
 
         for (final Comment comment : getComments()) {
-            comment.delete();
+            comment.delete(delOrder);
         }
-
         for (final Offer offer : getOffers()) {
-            offer.delete();
+            offer.delete(delOrder);
         }
-
         for (final Translation translation : getDescription().getTranslations()) {
-            translation.delete();
+            translation.delete(delOrder);
         }
-
-        for (final HighlightFeature hlFeature : HighlightFeatureManager.getAll()) {
-            if (hlFeature.getFeature().getId().equals(getId())) {
-                hlFeature.getDao().delete();
+        if (delOrder) {
+            for (final HighlightFeature hlFeature : HighlightFeatureManager.getAll()) {
+                if (hlFeature.getFeature().getId().equals(getId())) {
+                    hlFeature.getDao().delete();
+                }
             }
         }
     }
@@ -362,10 +402,12 @@ public final class FeatureImplementation extends Kudosable<DaoFeature> implement
         getDao().setFeatureState(FeatureState.DISCARDED);
 
         for (final Contribution contribution : getContributionsUnprotected()) {
-            contribution.cancel();
+            if (contribution.getState() == State.PENDING) {
+                contribution.cancel();
+            }
         }
         Offer selectedOffer = getSelectedOffer();
-        if (selectedOffer != null){
+        if (selectedOffer != null) {
             selectedOffer.cancelEverythingLeft();
         }
     }
@@ -544,6 +586,39 @@ public final class FeatureImplementation extends Kudosable<DaoFeature> implement
      */
     public void setMilestoneIsRejected() {
         setStateObject(getStateObject().eventMilestoneIsRejected());
+    }
+
+    @Override
+    public void setDescription(String newDescription, final Locale locale) throws UnauthorizedOperationException {
+        if (!canModify()) {
+            throw new UnauthorizedOperationException(Action.WRITE);
+        }
+
+        if (getDescription().getTranslation(locale) == null) {
+            throw new BadProgrammerException("Cannot modify a feature description for a non existing language. Should be a new translation");
+        }
+        getDao().setDescription(newDescription, locale);
+    }
+    
+    @Override
+    public void setTitle(String title, final Locale locale) throws UnauthorizedOperationException {
+        if (!canModify()) {
+            throw new UnauthorizedOperationException(Action.WRITE);
+        }
+        if (getDescription().getTranslation(locale) == null) {
+            throw new BadProgrammerException("Cannot modify a feature description for a non existing language. Should be a new translation");
+        }
+        
+        getDao().setTitle(title, locale);
+    }
+
+    @Override
+    public void setSoftware(Software software) throws UnauthorizedOperationException {
+        if (!canModify()) {
+            throw new UnauthorizedOperationException(Action.WRITE);
+        }
+
+        getDao().setSoftware(software.getDao());
     }
 
     // /////////////////////////////////////////////////////////////////////////////////////////
